tests and code optimizations
