+++ makefiles/virtual/namedparams.nc
@@ -1,0 +1,61 @@
+$#<_vmajor>
+$#<_vminor>
+$#<_line>
+g0x1y2z3
+g92x10y20z30
+G18G59.1G91G21
+F1000S500M3
+$G
+$#<_motion_mode>
+$#<_plane>
+$#<_ccomp>
+$#<_metric>
+$#<_imperial>
+$#<_absolute>
+$#<_incremental>
+$#<_inverse_time>
+$#<_units_per_minute>
+$#<_units_per_rev>
+$#<_coord_system>
+$#<_tool_offset>
+$#<_retract_r_plane>
+$#<_retract_old_z>
+$#<_spindle_rpm_mode>
+$#<_spindle_css_mode>
+$#<_ijk_absolute_mode>
+$#<_lathe_diameter_mode>
+$#<_lathe_radius_mode>
+$#<_spindle_on>
+$#<_spindle_cw>
+$#<_mist>
+$#<_flood>
+$#<_speed_override>
+$#<_feed_override>
+$#<_adaptive_feed>
+$#<_feed_hold>
+$#<_feed>
+$#<_rpm>
+$#<_x>
+$#<_y>
+$#<_z>
+$#<_a>
+$#<_b>
+$#<_c>
+$#<_u>
+$#<_v>
+$#<_w>
+$#<_abs_x>
+$#<_abs_y>
+$#<_abs_z>
+$#<_abs_a>
+$#<_abs_b>
+$#<_abs_c>
+$#<_current_tool>
+$#<_current_pocket>
+$#<_selected_tool>
+$#<_selected_pocket>
+$#<_value>
+$#<_value_returned>
+$#<_task>
+$#<_call_level>
+$#<_remap_level>
