+++ uCNC/src/hal/boards/avr/avr.ini
@@ -10,7 +10,7 @@ framework = arduino
 extra_scripts = 
 	${env.extra_scripts}
 	avr_compiler.py
-; debug_tool = simavr
+debug_tool = simavr
 debug_build_flags = -Og -g3 -ggdb3 -gdwarf-2
 build_flags = ${env.build_flags} -std=gnu99 -Wall -fdata-sections -ffunction-sections -fno-exceptions -Wl,--gc-sections -mcall-prologues -mrelax -flto -fno-fat-lto-objects -fno-tree-scev-cprop -Wl,--relax
 lib_deps = ${env.lib_deps}
