+++ uCNC/src/hal/mcus/virtual/virtual.ini
@@ -3,15 +3,15 @@
 ###############
 
 [env:EMULATOR_WINDOWS]
-platform = windows_x86
+platform = native
 framework =
 platform_packages = platformio/toolchain-gccmingw32
-build_src_filter = 
-  +<*>
-  +<../makefiles/virtual>
-  -<src/tinyusb>
-  -<src/hal/mcus/avr>
-  -<../makefiles/virtual/clib>
+; build_src_filter = 
+;   +<*>
+;   +<../makefiles/virtual/>
+;   -<src/tinyusb>
+;   -<src/hal/mcus/avr>
+;   -<../makefiles/virtual/clib>
 debug_build_flags = -Og -g3
 debug_init_cmds =
   run
